+++ include/color/operations/blend.hpp
@@ -18,11 +18,11 @@ namespace color::operations {
 namespace details {
 
 template <typename ColorType>
-static constexpr double get_scale_v() {
+static constexpr float get_scale_v() {
   if constexpr (std::is_integral_v<typename ColorType::value_type> && ColorType::scale() == 1) {
     return 255.0;
   } else {
-    return static_cast<double>(ColorType::scale());
+    return static_cast<float>(ColorType::scale());
   }
 }
 
@@ -43,9 +43,9 @@ constexpr Color1 blend(const Color1& c1, const Color2& c2, int ratio) {
 template <typename Color1, typename Color2>
 constexpr Color1 multiply(const Color1& c1, const Color2& c2) {
   auto c2_m = conversion::to_rgb<Color1>(c2);
-  constexpr double s = details::get_scale_v<Color1>();
+  constexpr float s = details::get_scale_v<Color1>();
   auto op = [s](auto v1, auto v2) {
-    return static_cast<typename Color1::value_type>((static_cast<double>(v1) * v2) / s);
+    return static_cast<typename Color1::value_type>((static_cast<float>(v1) * v2) / s);
   };
   return Color1{op(c1.r, c2_m.r), op(c1.g, c2_m.g), op(c1.b, c2_m.b)};
 }
@@ -53,7 +53,7 @@ constexpr Color1 multiply(const Color1& c1, const Color2& c2) {
 template <typename Color1, typename Color2>
 constexpr Color1 screen(const Color1& c1, const Color2& c2) {
   auto c2_m = conversion::to_rgb<Color1>(c2);
-  constexpr double s = details::get_scale_v<Color1>();
+  constexpr float s = details::get_scale_v<Color1>();
   auto op = [s](auto v1, auto v2) { return static_cast<typename Color1::value_type>(s - ((s - v1) * (s - v2) / s)); };
   return Color1{op(c1.r, c2_m.r), op(c1.g, c2_m.g), op(c1.b, c2_m.b)};
 }
@@ -61,11 +61,11 @@ constexpr Color1 screen(const Color1& c1, const Color2& c2) {
 template <typename Color1, typename Color2>
 constexpr Color1 overlay(const Color1& c1, const Color2& c2) {
   auto c2_m = conversion::to_rgb<Color1>(c2);
-  constexpr double s = details::get_scale_v<Color1>();
+  constexpr float s = details::get_scale_v<Color1>();
   auto op = [s](auto v1, auto v2) {
-    double f1 = static_cast<double>(v1);
-    double f2 = static_cast<double>(v2);
-    double res = (f1 < s / 2.0) ? (2.0 * f1 * f2 / s) : (s - 2.0 * (s - f1) * (s - f2) / s);
+    float f1 = static_cast<float>(v1);
+    float f2 = static_cast<float>(v2);
+    float res = (f1 < s / 2.0) ? (2.0 * f1 * f2 / s) : (s - 2.0 * (s - f1) * (s - f2) / s);
     return static_cast<typename Color1::value_type>(res);
   };
   return Color1{op(c1.r, c2_m.r), op(c1.g, c2_m.g), op(c1.b, c2_m.b)};
